% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.hapc.R
\name{kernel.hapc}
\alias{kernel.hapc}
\title{Kernel matrix computation for HAPC Models}
\usage{
kernel.hapc(X, max_degree, center = TRUE)
}
\arguments{
\item{X}{A numeric matrix containing the input features.}

\item{max_degree}{Integer specifying the maximum order of interaction terms (no more than max_degree - way interaction) for which
basis functions are generated. Default is 1 (no products) and might be increased until ncol(X).}

\item{center}{Logical indicating whether to center the basis functions before
processing. Default is TRUE.}
}
\value{
An \eqn{n \times n} kernel matrix \eqn{K = HH^T}, where n is the number of rows in X.
}
\description{
Computes the kernel matrix \eqn{K = HH^T} where H is the HAL (Highly Adaptive Lasso) 
design matrix generated by the hapc package.
}
\details{
This function does not explicitly build or store H in memory; instead, it applies 
the technique proposed in Schuler et al., Highly Adaptive Ridge, to work with H implicitly.
The ordering of the columns/rows of the output kernel matrix matches that of the 
kernel matrix computed from the HAL design matrix H obtained using the hal9001 package,
even though the design matrices differ by a column permutation. Consequently:
\itemize{
  \item The entries of the left singular vectors will be the same up to sign
  \item The eigenvalues will be exactly the same
}
}
\examples{
\dontrun{
# Create sample data
X <- matrix(rnorm(100 * 5), nrow = 100, ncol = 5)

# Compute kernel matrix with no interaction terms
K <- kernel.hapc(X, max_degree = 1)

# Compute kernel matrix with interaction terms up to 2-way and centering the columns of the underlying H. 
K <- kernel.hapc(X, max_degree = 2, center = TRUE)
}

}
